Room: New 'camera_rects' property
An array of rectangles defining camera boundaries.

Also adds two properties for controlling how the editor displays them.

In the _ready function, call _align_camera function and yield to it.

Player has a function called set_camera_limits_from_rect; it is called
by the level when it wants to adjust the camera limits.